make mapping extension a subpackage
