+++ src/vivarium_inputs/mapping_extension/__init__.py
@@ -1,0 +1,3 @@
+from .alternative_risk_factor import alternative_risk_factors
+from .healthcare_entity import healthcare_entities
+from .health_technology import health_technologies
