+++ src/vivarium_inputs/mapping_extension/health_technology.py
@@ -1,0 +1,10 @@
+from .health_technology_template import HealthTechnologies, HealthTechnology
+
+
+health_technologies = HealthTechnologies(
+    hypertension_drugs=HealthTechnology(
+        name='hypertension_drugs',
+        kind='health_technology',
+        gbd_id=None,
+    )
+)
